fix: resolve mypy type errors across vision, execution, and extraction modules
- Add ActionType union for cross-package Action compatibility in graph_executor
- Add WorkflowType union for dual Workflow support in graph_executor/traverser
- Fix BoundingBox type mismatches using Protocol pattern in ocr.py
- Replace inside() with with_region() in visibility, count assertions
- Fix OpenCV/numpy type annotations in screenshot, attributes, elements, typography
- Add assertions for optional dependency None access in OCR and SAM3 modules
- Update state.py to use ElementState/VisualStates instead of non-existent VisualState
- Various typing fixes across __init__ files and extraction modules

+++ src/qontinui/vision/verification/locators/base.py
@@ -40,7 +40,7 @@ class LocatorMatch:
     @property
     def area(self) -> int:
         """Get area of the match."""
-        return self.bounds.width * self.bounds.height
+        return int(self.bounds.width * self.bounds.height)
 
     def to_schema(self, locator_type: LocatorType, index: int = 0) -> VisionLocatorMatch:
         """Convert to schema model.
